[Refactor][Ops] read dtype at forward instead of taking it in op constructors (#1826)
## Summary

An op took `dtype` in its constructor and was then handed tensors
carrying one — two sources for one fact, and the constructor's copy is
the one that can be wrong. This makes the tensors the source: dtype is
read in `forward()`, and kernels are built there and cached per dtype.

- **Design** (`ops-design.md`, `ops-design-reference.md`
S12/S13/S16/S19): dtype is not a constructor parameter when the inputs
determine every output dtype. Seven ops are exempt because some output
dtype is an independent choice — two generative ops with no tensor
inputs, fp8 GEMM/BMM, DaCumsum, GQA prefill. The predicate is computable
from the manifest.
- **48 of 79 constructors migrated**: reduction 14, norm 6, MoE 9,
top-level 5, attention 14. Elementwise (24) and three attention ops that
select a kernel *class* from dtype are deferred to #1832 and #1833.
- **`is_hopper()` removed.** The library targets H200, so every branch
behind its false arm was unreachable. Six fallback kernels are now
reachable only through a `kernel_map` override; recorded in #1830.
- **One instance registry.** `elementwise/_base.py`, `rope.py` and
`dropout.py` each kept a private one, and 13 constructors overwrote
`dispatch_kernel`'s string key with `id(self)` — the `SymInt` hazard
`compile_boundary` documents was live across all 23 wrappers.
- **`_dtype_codegen`** emits the validator unrolled instead of looping
through `locals()`, which dynamo cannot trace; without it any op
validating dtypes in `forward` loses `fullgraph`.
- **`Op.autotune`** searched for `Kernel`-typed attributes and so found
nothing once kernels moved into caches. It now walks them.

## Test plan

- [x] Per-family suites green; `-m smoke` green across ops
- [x] `validate_manifest.py --strict` clean
- [x] New `test_multi_dtype_instance.py`: eight cases over reduction,
norm, attention decode, MoE and a top-level op, each feeding one
instance two dtypes; plus a cold `fullgraph=True` call on a
dtype-inferred op
- [ ] Benchmarks: no kernel selection changed, so no perf delta is
expected

## Notes for review

- Behaviour change: an invalid dtype or out-of-range scalar now raises
at first use rather than at construction.
- Manifest edits are confined to what the code change forces:
`source.kernel_map` retargeting after the Hopper cleanup, three `dtype`
params the ops no longer accept, and a declared output dtype GQA prefill
was already selecting.
- The validator compared a `torch.dtype` default against its YAML
spelling with `!=`, so no entry could declare one. Fixed with a test.

+++ benchmarks/ops/attention/bench_deepseek_dsa_decode.py
@@ -81,7 +81,7 @@ def test_dsa_decode_bench(batch: int, heads: int, seq_len_q: int, seq_len_kv: in
 
     op = DeepSeekSparseAttentionDecodeWithKVCacheFwdOp(
         batch, heads, seq_len_q, seq_len_kv, dim, dim_tail, topk, stride_kv, heads_kv,
-        q_start_index_s, sm_scale=sm_scale, dtype=dtype, tune=tune)
+        q_start_index_s, sm_scale=sm_scale, tune=tune)
     bm = ManifestBenchmark(_OP_NAME, op, test)
     result = bm.profile(op, *inputs)
     BenchmarkReport.record(op, locals(), result, tag="tileops")
