+++ tests/ops/test_normalization_alignment.py
@@ -17,7 +17,7 @@ import torch
 def test_rms_norm_accepts_normalized_shape() -> None:
     from tileops.ops.norm.rms_norm import RMSNormFwdOp
 
-    op = RMSNormFwdOp(normalized_shape=(4096,), eps=None, dtype=torch.float16)
+    op = RMSNormFwdOp(normalized_shape=(4096,), eps=None)
     assert op.N == 4096
     assert op.normalized_shape == (4096,)
 
@@ -26,7 +26,7 @@ def test_rms_norm_accepts_normalized_shape() -> None:
 def test_layer_norm_accepts_normalized_shape() -> None:
     from tileops.ops.norm.layer_norm import LayerNormFwdOp
 
-    op = LayerNormFwdOp(normalized_shape=[4096], dtype=torch.float16)
+    op = LayerNormFwdOp(normalized_shape=[4096])
     assert op.N == 4096
     assert op.normalized_shape == (4096,)
 
@@ -41,7 +41,7 @@ def test_rms_norm_accepts_tuple_normalized_shape_runtime() -> None:
 
     from tileops.ops.norm.rms_norm import RMSNormFwdOp
 
-    op = RMSNormFwdOp(normalized_shape=(2, 3), dtype=torch.float16)
+    op = RMSNormFwdOp(normalized_shape=(2, 3))
     assert op.N == 6
     assert op.normalized_shape == (2, 3)
     x = torch.randn(4, 2, 3, dtype=torch.float16, device="cuda")
@@ -59,7 +59,7 @@ def test_layer_norm_accepts_tuple_normalized_shape_runtime() -> None:
 
     from tileops.ops.norm.layer_norm import LayerNormFwdOp
 
-    op = LayerNormFwdOp(normalized_shape=(2, 3), dtype=torch.float16)
+    op = LayerNormFwdOp(normalized_shape=(2, 3))
     assert op.N == 6
     assert op.normalized_shape == (2, 3)
     x = torch.randn(4, 2, 3, dtype=torch.float16, device="cuda")
